requirements.txt
# Core Flask dependencies
Flask>=3.0.0
Werkzeug>=3.0.0
Flask-Login>=0.6.3
Flask-SQLAlchemy>=3.1.1
Flask-Migrate>=4.0.7
Flask-WTF>=1.2.1
Flask-Mail>=0.10.0
Flask-Babel>=4.0.0
Flask-Session>=0.8.0
Flask-DebugToolbar>=0.16.0
Flask-Alembic>=3.1.1

# Database
SQLAlchemy>=2.0.36
alembic>=1.13.0
greenlet>=3.0.0
mysqlclient>=2.2.0

# Forms and Validation
WTForms>=3.1.2
email-validator>=2.1.1

# Security
PyJWT>=2.8.0
cryptography>=41.0.3
itsdangerous>=2.2.0

# Data Processing
pandas>=2.2.0
numpy>=1.21.0,<2.0.0
openpyxl>=3.1.2
XlsxWriter>=3.2.0

# Utilities
python-dateutil>=2.8.2
pytz>=2024.1
Jinja2>=3.1.4
MarkupSafe>=3.0.2
click>=8.1.7
blinker>=1.8.2

# Development tools
python-dotenv>=1.0.0